Vector<int16_t, 2>::getLength()  cannot return a float, and methode already implemented above, most likely  Vector<float, 2>::getLength()  wanted

+++ src/xpcc/math/geometry/vector2.cpp
@@ -52,7 +52,7 @@ namespace xpcc
 	
 	template<>
 	float
-	Vector<int16_t, 2>::getLength() const
+	Vector<float, 2>::getLength() const
 	{
 		return sqrtf(getLengthSquared());
 	}
